requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.1
django-cors-middleware==1.3.1
django-extra-fields==0.8
djangorestframework==3.4.6
Pillow==3.4.1
python-dateutil==2.5.3
